Fix upgrade version of Morph
We wanted to upgrade Morph to use latest master, and because of
an  human error of a Friday evening we used the sha1 of master of
definitions.git instead of master of morph.git in the commit
c3586f5c01937e19ed1b4a4ea7f195b5045e9201.

Given that I gave a +2 for upgrading Morph to latest master, I'm
merging this change without sending it for review.