gerrit: Move system and stratum .morph files into strata/ and systems/
This is so others can reuse them without having to reuse the other stuff
in baserock_gerrit. Most likely people who are setting up a Gerrit with
Baserock will make a copy of the baserock_gerrit/ folder, rather than
reusing it directly. If they copied the .morph files then they'd miss
out on improvements made to those files in subsequent commits to
infrastructure.git.

Such users will still miss out on improvements to the Ansible modules --
hopefully we can solve that in a nice way in future, too.