Make 'hosts' a symlink to baserock_hosts
The idea is to make it easier for people to fork infrastructure.git and
use it for their own infrastructure. They'll need to totally change
'hosts' to point to their own systems, and this would lead to merge
conflicts every time they tried to pull in 'master' of
infrastructure.git.