Use HTTPS for all infrastructure.
This is implemented with the HAProxy frontend doing 'SSL termination'.
So internal traffic between the frontend_haproxy instance and the
various machines serving content is unencrypted HTTP as before, but all
traffic that goes over the public internet is encrypted now.

Note that storyboard.baserock.org is not behind HAProxy, and currently
uses a different, self-signed certificate.

Change-Id: I9140def605fe26c9c613066fa6524e3cf817f97c