gerrit: Allow more people to push to refs/heads/*
The intention was always that Mergers would be able to push to anything
in refs/heads/* (on the assumption that we can trust each other not to
bypass the accepted review processes). Seems I never actually
implemented that.

Also, allow forgeAuthor so Mergers can push branches with commits made
by other people (without this, Gerrit complains if the author of any of
the commits doesn't match the email address of the person pushing).

Change-Id: Id60659b51f08bfcec9af2f8681a4faf958301bd0