gerrit: Don't let anyone force-push to 'master'
Previously, Mirroring Tools (Lorry) could force-push anywhere. This is
so that personal branches are kept in sync between git.baserock.org and
gerrit.baserock.org. Since users may force-push to their personal
branches, it's necessary to allow force-pushes. But nobody should force
'master', and I have a feeling that this is causing an issue we are
seeing where Gerrit says that it has merged something, but there is no
sign of the merge in 'master'.

Change-Id: I80bc4eace46470ffa7f3da185fcc1c1f228cda71