Create kmod symlinks in sbin
iSCSI kernel modules and open-iscsi expect modprobe and the rest
of the module kernel utils to be in /sbin.
This patch create the links in /sbin instead of /bin if /sbin
is not a symlink of /usr/bin; if it is a symlink it will create
the links in /usr/bin.
Also it installs kmod in "$PREFIX" instead of /

Change-Id: I89226d0ee7956e294caa66d60101a35daeccf9a1