Revert "Add a compatibility symlink for /lib/ld-linux.so.3 on ARMv7 hard-float"
This reverts commit a51a8ee9fe2c691a4b2d73fe06c1fde1d18f1153.

This symlink was a bad idea. We found that if the stage1-gcc chunk is
built on an ARM system with EGLIBC 2.15, then run on a system with GLIBC
2.20, the GCC binaries segfault on startup, inside the _dl_addr()
function. This is likely the result of an ABI compatibility between
EGLIBC 2.15 and GLIBC 2.20.

Morph still has a big hole in its caching algorithm whereby 'bootstrap'
mode chunks built against incompatible libc versions get the same
cache-key, and cause random breakages. We either need to fix this, or
ensure that the reference bootstrap is entirely statically linked, as
soon as possible.

Change-Id: I400d293dc8022ca6ab5a0ebda62ce378e9bf3317