Configure and start an iSCSI initiator on all required machines
The OpenStack Ironic Conductor service requires that an iSCSI initiator
is running on the same local machine.  This commit configures and sets
an iSCSI initiator to run on the controller node on a two-node and
three-node OpenStack system.  For a one-node system, this service was
already configured as it is required by OpenStack Cinder.

Change-Id: I46291d9cd1a31e3ff91888401d45ab0dc67e3677