pxeboot: add support for PXEBOOT_TARGET_INTERFACE
This is to set the name of the interface of the target to pxeboot from.
Some targets with more than one interface try to get the rootfs from a
different interface than the interface from where the pxeboot server is
reachable. Using this variable, the kernel arguments will be filled
to include the device.

Change-Id: I022bf6b17a66d63f7f9b8830e72ca4020158a80a