openstack: improve services' logging
The OpenStack services write their log to stderr even when use_syslog
was set to True in the configuration file.  This is then captured by
journald, resulting in duplicated entries on the log.  Set
StandardOutput=null and StandardError=null in the unit files to prevent
this.

As the services log is being dumped on journald, there is no need to
duplicate the log in separate log files, so remove --log-file from the
unit files as well.

Last but not least, enable writing to the syslog for Ironic and
Keystone.

Change-Id: I7808e8116fc92ecb807f700fd94c9ecf4ee13be2