Move GLib and gobject-introspection to -core
Morph now depend on pygobject (to access ostree features).
Also ostree depends on libsoup.
Both libsoup and pygobject depends on GLib.
morph is in cross-* systems, so we move GLib
to the core stratum to not have to build the entire foundation
stratum in the cross-* systems.
Also the change makes sense as GLib is a dependency of strata
that not necessarily depend on foundation (like
multimedia-gstreamer)

Change-Id: I9d1527b5057c71704ef5a0fe3d3e5f348f2908e5