openstack: ensure that /var/run/openvswitch exists
This directory is created by the setup script, but as /var/run is a
symbolic link to /run, which is a tmpfs mounting point, the directory
will not persist across reboots.  Hence, ensure that the directory
exists before starting the services that require it.

Note that we can't get away by adding the openvswitch-setup.service to
the Wants of openvswitch.service and openvswitch-db-server.services, as
those are enabled and started by the Ansible setup script, and so would
create a deadlock.

Also, remove non-existent service from the After section.

Change-Id: Ic5c16f89a8e71e4ad42f468159d6e604a5ee8e47