Ensure Masons can /only/ fetch sources from git.baserock.org
This is done by using 'internal-only' instead of 'default' as the
security group. I've updated the README to note this. To make Mason
work in the absence of DNS we also have to change the 'trove-host'
setting to be the actual IP of git.baserock.org.

The idea is to enforce the policy that the Baserock reference system
definitions can only use 'baserock:' and 'upstream:' keyed URLs.

Change-Id: I114fc89a707f6f626e4b758426558f48e5fafb73