Mason workspace should be in /srv/distbuild
The mason.sh script does all its work in a directory called 'ws',
created in the current directory of wherever it is run from. When run as
a system service it creates /ws in the root filesystem.

The work Mason does involves creating one or more disk images, and these
can require a lot of disk space and IO bandwidth. The root filesystem of
a machine isn't a good place to do this sort of thing: running out of
diskspace there will completely break the system, and it may be stored
on an SD card or eMMC.

There is an established pattern in deploying Baserock 'distbuild'
systems to have an external hard disk or volume mounted at
/srv/distbuild, so that running builds and caching built artifacts
doesn't take place on the root filesystem. It makes sense to move the
Mason workspace in there too.

Change-Id: I39f53373a7411c8816e886886c8af40e08f4782e