systems: Add coreutils-common to all devel systems
This is not generally included in systems because it contains GPLv3
components, and we support GPLv3-less systems.

However, devel systems aren't among the systems we need to ensure
contain no GPLv3 components, and they are required for building some
components, and building components outside of morph is a defined
use-case for devel systems.

Change-Id: I8475af70bda1b26e04adebce8897f272a1aa18bf