Add systems/build-system-armv5l-openbmc-aspeed.morph
After cross-building an armv5l system, you should build this system,
and use this system to build future armv5l-openbmc-aspeed systems.
This is because the cross-built system is fragile and does not allow
you to deploy a system locally, as the cross-built system does not
contain btrfs-progs. This system does

Change-Id: Ibc73f08d5c24b8ee4d100cdb3c499786a319293e