Update ref of librabitmq to a fixed version
The previous version had its submodules pointing to
git.baserock.org. The right thing to do is to point to the
Trove using the 'upstream:' alias, so that definitions don't
use git.baserock.org.

This is needed if you want to build using a local trove, with
all the sources (downstream trove) and you don't have access
to git.baserock.org.

Change-Id: Id8b26064d0afd66a0b1db6a18c4b69e02085299d