libvirt.morph: Don't autostart the default network
Remove the symlink that enables the default network to autostart.
It's  helpful to have a default network, but starting it
automatically may create some confusion to the network if you
didn't want it running, and didn't know it was there.

I found that this network was running after configuring a OpenStack
sub-network using the same range of IP addresses and suffering
connectivity issues.

Change-Id: I0d0c399236ba06547e20c6433f115633a97b01b9