Move erlang-sd_notify to openstack-service stratum
erlang-sd_notify is required only, at the moment, by rabbitmq server
to be able to send notifications to systemd, so makes sense for it
to belongs only to openstack-service stratum making the erlang
stratum minimal. This movement makes erlang stratum not depends on
systemd (because erlang-sd_notify needs systemd headers to compile),
defo it can depends only in core stratum.

Change-Id: Ie6fbb58502336d21fa882bae09860a56d4c14a60