Cinder: Wipe only the first 50 MiB of the Volumes when deleting
This will wipe the partition table of the Volume making the
process of deleting a volume faster.
Note that before when deleting a volume it was filled by zeros wiping
all the contents on it no matter the size of the volume. This
process is very long and can cause performance issues to the
the users if some volumes are deleting a the same time.

Change-Id: I701810758683b715c98e10ac6a923c48afc519c4