Improve ntpd systemd unit and configuration
Start ntpd with the `-N` and `-g` options, as recommended in the man
page.  `-N` will, to the extent permitted by the operating system, run
the ntpd at the highest priority.  `-g` will make ntpd to update the
time even if the offset with the current time is greater than 1000s.

Add 'limited' to the restrict options to avoid the following warning in
the log: 'restrict ::: KOD does nothing without LIMITED'.

Change-Id: I23e264533ef27af7a0504e8cd5b69c7e01843728